3d-viewer-tool-python/requirements.txt
numpy==1.14.0
pkg-resources==0.0.0
PyGLM==0.2.1a1
PyOpenGL==3.1.1a1
transforms3d==0.3.1
